# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/config": {
            "post": {
                "description": "The following API endpoint can be used to Agent configure.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent",
                    "connector"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth header",
                        "name": "GWOS-APP-NAME",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Auth header",
                        "name": "GWOS-API-TOKEN",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {},
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/events": {
            "post": {
                "description": "The following API endpoint can be used to send Alerts to Foundation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alert",
                    "event"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth header",
                        "name": "GWOS-APP-NAME",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Auth header",
                        "name": "GWOS-API-TOKEN",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {},
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/events-ack": {
            "post": {
                "description": "The following API endpoint can be used to send Alerts to Foundation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alert",
                    "event"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth header",
                        "name": "GWOS-APP-NAME",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Auth header",
                        "name": "GWOS-API-TOKEN",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {},
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/events-unack": {
            "post": {
                "description": "The following API endpoint can be used to send Alerts to Foundation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alert",
                    "event"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth header",
                        "name": "GWOS-APP-NAME",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Auth header",
                        "name": "GWOS-API-TOKEN",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {},
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/metrics": {
            "get": {
                "description": "The following API endpoint can be used to get list of metrics from the server.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metric"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth header",
                        "name": "GWOS-APP-NAME",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Auth header",
                        "name": "GWOS-API-TOKEN",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.AgentStatus"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/start": {
            "post": {
                "description": "The following API endpoint can be used to start NATS dispatcher.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent",
                    "connector"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth header",
                        "name": "GWOS-APP-NAME",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Auth header",
                        "name": "GWOS-API-TOKEN",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.ConnectorStatusDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/stats": {
            "get": {
                "description": "The following API endpoint can be used to get TCG statistics.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent",
                    "connector"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth header",
                        "name": "gwos-app-name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Auth header",
                        "name": "GWOS-API-TOKEN",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.AgentStats"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/status": {
            "get": {
                "description": "The following API endpoint can be used to get TCG status.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent",
                    "connector"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth header",
                        "name": "GWOS-APP-NAME",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Auth header",
                        "name": "GWOS-API-TOKEN",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.ConnectorStatusDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/stop": {
            "post": {
                "description": "The following API endpoint can be used to stop NATS dispatcher.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent",
                    "connector"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth header",
                        "name": "GWOS-APP-NAME",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Auth header",
                        "name": "GWOS-API-TOKEN",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.ConnectorStatusDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "milliseconds.MillisecondTimestamp": {
            "type": "object"
        },
        "services.AgentStats": {
            "type": "object",
            "properties": {
                "agentID": {
                    "type": "string"
                },
                "appType": {
                    "type": "string"
                },
                "bytesSent": {
                    "type": "integer"
                },
                "executionTimeInventory": {
                    "type": "string"
                },
                "executionTimeMetrics": {
                    "type": "string"
                },
                "lastAlertRun": {
                    "type": "object",
                    "$ref": "#/definitions/milliseconds.MillisecondTimestamp"
                },
                "lastErrors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "lastInventoryRun": {
                    "type": "object",
                    "$ref": "#/definitions/milliseconds.MillisecondTimestamp"
                },
                "lastMetricsRun": {
                    "type": "object",
                    "$ref": "#/definitions/milliseconds.MillisecondTimestamp"
                },
                "messagesSent": {
                    "type": "integer"
                },
                "metricsSent": {
                    "type": "integer"
                },
                "upSince": {
                    "type": "object",
                    "$ref": "#/definitions/milliseconds.MillisecondTimestamp"
                }
            }
        },
        "services.AgentStatus": {
            "type": "object",
            "properties": {
                "controller": {
                    "type": "string"
                },
                "ctrl": {
                    "type": "object",
                    "$ref": "#/definitions/services.CtrlAction"
                },
                "nats": {
                    "type": "string"
                },
                "transport": {
                    "type": "string"
                }
            }
        },
        "services.ConnectorStatusDTO": {
            "type": "object",
            "properties": {
                "connectorStatus": {
                    "type": "string"
                },
                "jobId": {
                    "type": "integer"
                }
            }
        }
    }
}